# Speed up jump-graph generation