--- a/.github/workflows/python-package.yml
+++ b/.github/workflows/python-package.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Lint with ruff - syntax errors & undefined names
         run: |
           # stop the build if there are Python syntax errors or undefined names
-          ruff --select=E9,F63,F7,F82 --preview PyRoute
+          ruff check --select=E9,F63,F7,F82 --preview PyRoute
       - name: Lint with ruff - configured options
         run: |
           # now run configured options