--- a/PyRoute/Calculation/RouteCalculation.py
+++ b/PyRoute/Calculation/RouteCalculation.py
@@ -129,8 +129,8 @@ class RouteCalculation(object):
         code1 = star1.tradeCode
         code2 = star2.tradeCode
         if code1.agricultural or code2.agricultural:
-            if (code1.agricultural and (code2.nonagricultural or code2.extreme)) or \
-               ((code1.nonagricultural or code1.extreme) and code2.agricultural):
+            if (code1.agricultural and code2.needs_agricultural) or \
+               (code1.needs_agricultural and code2.agricultural):
                 btn += 1
         if code1.industrial or code2.industrial:
             if (code1.nonindustrial and code2.industrial) or (code2.nonindustrial and code1.industrial):