--- a/PyRoute/Calculation/TradeCalculation.py
+++ b/PyRoute/Calculation/TradeCalculation.py
@@ -60,6 +60,7 @@ class TradeCalculation(RouteCalculation):
         # this value are ignored. Set lower to have more routes calculated, but
         # may not have have an impact on the overall trade flows.
         self.min_btn = min_btn
+        self.min_route_wtn = (min_btn - 1) // 2  # In light of minimum btn, what is smallest WTN that _can_ meet it?
 
         # Minimum WTN to process routes for
         self.min_wtn = route_btn
@@ -92,6 +93,14 @@ class TradeCalculation(RouteCalculation):
         # by the time we've _reached_ here, we're assuming generate_base_routes() has handled the unilateral filtering
         # - in this case, red/forbidden zones and barren systems - so only bilateral filtering remains.
         # TODO: Bilateral filtering
+        # This would ordinarily be a unilateral filter, but, for hysterical raisins, route and edge filtering are
+        # convolved.  Rather than untangle that, filter out routes with at least one endpoint too small to support the
+        # minimum WTN route here.
+        if self.min_route_wtn > star.wtn or self.min_route_wtn > neighbor.wtn:
+            # Don't filter if, despite the route being too small, it's within the max jump range.  Such stars can still
+            # have trade routes flowing _through_ them, just not _from_ or _to_ them.
+            if self.galaxy.max_jump_range < star.distance(neighbor):
+                return True
         return False
 
     def base_range_routes(self, star, neighbor):
@@ -101,6 +110,7 @@ class TradeCalculation(RouteCalculation):
         # if there there isn't enough trade to warrant a trade check
         if dist > max(self.galaxy.max_jump_range, max_dist):
             return None
+
         if dist <= max_dist:
             # Only bother getting btn if the route is inside max length
             btn = self.get_btn(star, neighbor, dist)
@@ -110,11 +120,6 @@ class TradeCalculation(RouteCalculation):
                                             btn=btn,
                                             passenger_btn=passBTN)
         return dist
-
-    def _raw_ranges(self):
-        ranges = super(RouteCalculation)._raw_ranges()
-
-        return ranges
 
     def generate_routes(self):
         """