--- a/PyRoute/Calculation/TradeCalculation.py
+++ b/PyRoute/Calculation/TradeCalculation.py
@@ -108,7 +108,7 @@ class TradeCalculation(RouteCalculation):
         max_dist = self._max_dist(star.wtn, neighbor.wtn)
         # add all the stars in the BTN range, but skip this pair
         # if there there isn't enough trade to warrant a trade check
-        if dist > max(self.galaxy.max_jump_range, max_dist):
+        if dist > max_dist and dist > self.galaxy.max_jump_range:
             return None
 
         if dist <= max_dist: