--- a/PyRoute/TradeCodes.py
+++ b/PyRoute/TradeCodes.py
@@ -405,15 +405,15 @@ class TradeCodes(object):
             return [code if len(code) > 6 else 'C:' + sector_name[0:4] + '-' + code[2:]
                     for code in self.owned if code.startswith('C:')]
 
-    @property
+    @functools.cached_property
     def homeworld(self):
         return sorted(self.homeworld_list)
 
-    @property
+    @functools.cached_property
     def sophonts(self):
         return sorted(self.sophont_list)
 
-    @property
+    @functools.cached_property
     def rich(self):
         return 'Ri' in self.codeset
 
@@ -429,75 +429,75 @@ class TradeCodes(object):
     def needs_agricultural(self):
         return self.nonagricultural or self.extreme
 
-    @property
+    @functools.cached_property
     def poor(self):
         return 'Po' in self.codeset
 
-    @property
+    @functools.cached_property
     def nonagricultural(self):
         return 'Na' in self.codeset
 
-    @property
+    @functools.cached_property
     def barren(self):
         return 'Ba' in self.codeset or 'Di' in self.codeset
 
-    @property
+    @functools.cached_property
     def low(self):
         return 'Lo' in self.codeset
 
-    @property
+    @functools.cached_property
     def nonindustrial(self):
         return 'Ni' in self.codeset
 
-    @property
+    @functools.cached_property
     def high(self):
         return 'Hi' in self.codeset
 
-    @property
+    @functools.cached_property
     def asteroid(self):
         return 'As' in self.codeset
 
-    @property
+    @functools.cached_property
     def desert(self):
         return 'De' in self.codeset
 
-    @property
+    @functools.cached_property
     def fluid(self):
         return 'Fl' in self.codeset
 
-    @property
+    @functools.cached_property
     def vacuum(self):
         return 'Va' in self.codeset and 'As' not in self.codeset
 
-    @property
+    @functools.cached_property
     def waterworld(self):
         return 'Wa' in self.codeset or 'Oc' in self.codeset
 
-    @property
+    @functools.cached_property
     def extreme(self):
         return len(self.ex_codes & set(self.codeset)) > 0
 
-    @property
+    @functools.cached_property
     def capital(self):
         return 'Cp' in self.dcode or 'Cx' in self.dcode or 'Cs' in self.dcode
 
-    @property
+    @functools.cached_property
     def subsector_capital(self):
         return 'Cp' in self.dcode
 
-    @property
+    @functools.cached_property
     def sector_capital(self):
         return 'Cs' in self.dcode
 
-    @property
+    @functools.cached_property
     def other_capital(self):
         return 'Cx' in self.dcode
 
-    @property
+    @functools.cached_property
     def research_station(self):
         return set(self.research.keys()).intersection(self.dcode)
 
-    @property
+    @functools.cached_property
     def research_station_char(self):
         stations = self.research_station
         if len(stations) == 1:
@@ -506,11 +506,11 @@ class TradeCodes(object):
         else:
             return None
 
-    @property
+    @functools.cached_property
     def pcode_color(self):
         return self.pcolor.get(self.pcode, '#44ff44')
 
-    @property
+    @functools.cached_property
     def low_per_capita_gwp(self):
         return self.extreme or self.poor or self.nonindustrial or self.low